♻️ Refactor: 모든챌린지페이지 리팩토링
