+++ src/components/CategoryFilter.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import s from './styles/CategoryFilter.module.scss';
+import s from './style/CategoryFilter.module.scss';
 
 const categories = ['전체', '운동', '식습관', '생활', '기타'];
