+++ src/pages/exploreChallengePage/ExploreChallengePage.jsx
@@ -1,6 +1,7 @@
 import React, { useState } from 'react';
+import s from '././ExploreChallengePage.module.scss';
 import Header from './components/Header.jsx';
-import CategoryFilter from './components/CategoryFilter.jsx';
+import CategoryFilter from '../../components/CategoryFilter.jsx';
 import AllChallenge from './components/AllChallengeBig.jsx';
 import dummyData from './datas/AllChallengeDummy.json';
 
@@ -14,7 +15,7 @@ const ExploreChallengePage = () => {
       : dummyData.items.filter((c) => c.category.name === selectedCategory);
 
   return (
-    <div>
+    <div className={s.exploreChallengePageContainer}>
       <Header />
       <CategoryFilter selectedCategory={selectedCategory} onSelect={setSelectedCategory} />
       <AllChallenge challenges={filteredItems} />
