+++ src/pages/mainPage/components/styles/AllChallenge.module.scss
@@ -17,6 +17,7 @@
     color: $gray-500;
     background: none;
     border: none;
+    cursor: pointer;
   }
   .moreIcon {
     width: 10px;
