+++ src/pages/mainPage/components/styles/Header.module.scss
@@ -19,6 +19,7 @@
       text-decoration: underline;
       text-decoration-color: #fcfcfc;
       text-underline-offset: 2px;
+      cursor: pointer;
     }
   }
   .headerBottom {
